Fix 64-bit compile.
BUG=
R=jarin@chromium.org

Review URL: https://codereview.chromium.org/397373006

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@22472 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/mark-compact.cc
@@ -3974,14 +3974,14 @@ static inline Address StartOfLiveObject(Address block_address, uint32_t cell) {
 
 // Force instantiation of templatized SweepConservatively method for
 // SWEEP_ON_MAIN_THREAD mode.
-template intptr_t MarkCompactCollector::
+template int MarkCompactCollector::
     SweepConservatively<MarkCompactCollector::SWEEP_ON_MAIN_THREAD>(
         PagedSpace*, FreeList*, Page*);
 
 
 // Force instantiation of templatized SweepConservatively method for
 // SWEEP_IN_PARALLEL mode.
-template intptr_t MarkCompactCollector::
+template int MarkCompactCollector::
     SweepConservatively<MarkCompactCollector::SWEEP_IN_PARALLEL>(
         PagedSpace*, FreeList*, Page*);
